backend/HackChennai/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-phonenumber-field==5.0.0
djangorestframework==3.11.1
phonenumbers==8.12.9
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
